pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coach-agents"
version = "0.1.0"
description = "Orchestrate personal coaching AI agents powered by Claude Code CLI and Telegram."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "product-teammate" }]
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
    "python-telegram-bot>=21",
    "loguru>=0.7",
    "aiosqlite>=0.20",
    "jsonschema>=4.22",
    "apscheduler>=3.10",
    "httpx>=0.27",
    "markdownify>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "black>=24.0",
]

[project.scripts]
coach = "coach_cli.__main__:app"

[tool.setuptools.packages.find]
include = ["brains*", "channels*", "runtime*", "coach_cli*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
